Create Bid functionalities
Possible for a user to send a bid in the chat between them and the
seller. The seller can accept or reject the incoming bid. The buyer can
cancel the bid if it is not yet accepted.
When a bid is accepted, the listingStatus is set to SOLD. The listing
cannot be activated after being marked as SOLD.